+++ _legacy.php
@@ -73,6 +73,7 @@ $__autoload = [
     'formSelect'                => $src . '/Html/Form/Select.php',
     'formOptgroup'              => $src . '/Html/Form/Optgroup.php',
     'formOption'                => $src . '/Html/Form/Option.php',
+    'pager'                     => $src . '/Html/Pager.php',
     'imageMeta'                 => $src . '/Image/ImageMeta.php',
     'imageTools'                => $src . '/Image/ImageTools.php',
     'mail'                      => $src . '/Mail/Mail.php',
@@ -97,7 +98,6 @@ $__autoload = [
     'xmlrpcClientMulticall'     => $src . '/Network/Xmlrpc/ClientMulticall.php',
     'xmlrpcServer'              => $src . '/Network/Xmlrpc/Server.php',
     'xmlrpcIntrospectionServer' => $src . '/Network/Xmlrpc/IntrospectionServer.php',
-    'pager'                     => $src . '/Pager/Pager.php',
     'restServer'                => $src . '/Rest/Server.php',
     'sessionDB'                 => $src . '/Session/SessionDb.php',
     'template'                  => $src . '/Template/Template.php',
